requirements.txt
requests
pandas
python-dotenv
psycopg2
streamlit
plotly-express